AP-1136 rebase on master
